Show device model on QR login success step
Step 3 (success step) was rendering buildQRLoginDeviceSubline, which
omits the device model (it only shows OS + app version) because the
standalone QRLoginConsumedPanel already names the model in its headline.
The modal flow has no such headline, so the model — the most
recognizable field for spotting a wrong-device scan — was lost.

Promote the number-match step's model-inclusive line builder into the
shared qrLoginDeviceCopy as buildQRLoginDeviceLine (model · OS version ·
App version) and use it from both step 2 and step 3, so the success step
shows the same device line as the number-match step. The standalone
panel keeps using buildQRLoginDeviceSubline (model-less) to avoid
duplicating the model it already shows in its headline.
